requirements_Version3.txt
python-telegram-bot==20.6
python-dotenv==1.0.1
pytz==2024.1